+++ R/module-auth.R
@@ -8,6 +8,8 @@
 #' @param tags_top A \code{tags (div, img, ...)} to be displayed on top of the authentication module.
 #' @param tags_bottom A \code{tags (div, img, ...)} to be displayed on bottom of the authentication module.
 #' @param background A optionnal \code{css} for authentication background. See example.
+#' @param choose_language \code{logical/character}. Add language selection on top ? TRUE for all supported languages
+#' or a vector of possibilities like \code{c("fr", "en")}. If enabled, \code{input$shinymanager_language} is created
 #' @param ... : Used for old version compatibility.
 #' 
 #' 
@@ -16,16 +18,19 @@
 #' @name module-authentication
 #'
 #' @importFrom htmltools tagList tags singleton
-#' @importFrom shiny NS fluidRow column textInput passwordInput actionButton
+#' @importFrom shiny NS fluidRow column textInput passwordInput actionButton uiOutput 
 #'
 #' @example examples/module-auth.R
 auth_ui <- function(id, status = "primary", tags_top = NULL, 
-                    tags_bottom = NULL, background = NULL, ...) {
-
+                    tags_bottom = NULL, background = NULL, 
+                    choose_language = NULL, lan = NULL, ...) {
+  
   ns <- NS(id)
-
-  lan <- use_language()
-
+  
+  if(is.null(lan)){
+    lan <- use_language()
+  }
+  
   # patch / message changing tag_img & tag_div
   deprecated <- list(...)
   if("tag_img" %in% names(deprecated)){
@@ -55,10 +60,44 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
             class = paste0("panel panel-", status),
             tags$div(
               class = "panel-body",
+              {
+                choices = lan$get_language()
+                if(is.logical(choose_language) && choose_language){
+                  choices = lan$get_language_registered()
+                } else if(is.character(choose_language)){
+                  choices = unique(c(intersect(choose_language, lan$get_language_registered()), lan$get_language()))
+                }
+                
+                  selected = ifelse(lan$get_language() %in% choices, 
+                                    lan$get_language(),
+                                    choices[1])
+                  if(length(choices) == 1){
+                    style = "display:none"
+                  } else {
+                    style = "margin-bottom:-50px;"
+                  }
+                  tags$div(style = style,
+                           fluidRow(
+                             column(width = 3, offset = 6, uiOutput(ns("label_language"))),
+                             column(3,
+                                    tags$div(
+                                      style = "text-align: left; font-size: 12px;",
+                                      selectInput(
+                                        inputId = ns("language"),
+                                        label = NULL,
+                                        choices = choices,
+                                        selected = selected,
+                                        width = "100%"
+                                      )
+                                    )
+                             )
+                           )
+                  )
+              },
               tags$div(
                 style = "text-align: center;",
                 if (!is.null(tags_top)) tags_top,
-                tags$h3(lan$get("Please authenticate"))
+                tags$h3(lan$get("Please authenticate"), id = ns("shinymanager-auth-head"))
               ),
               tags$br(),
               textInput(
@@ -83,7 +122,8 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
                 sprintf("bindEnter('%s');", ns(""))
               ),
               tags$div(id = ns("result_auth")),
-              if (!is.null(tags_bottom)) tags$hr(), tags_bottom
+              if (!is.null(tags_bottom)) tags$div(style = "margin-top:-10px;", tags$hr()), tags_bottom,
+              uiOutput(ns("update_shinymanager_language"))
             )
           )
         )
@@ -100,7 +140,8 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
 #'  Must return \code{TRUE} or \code{FALSE}.
 #'  To use additionnals arguments, set them with \code{purrr::partial} (see examples).
 #' @param use_token Add a token in the URL to check authentication. Should not be used directly.
-#'
+#' @param lan An langauge object. Should not be used directly.
+#' 
 #' @export
 #'
 #' @rdname module-authentication
@@ -113,19 +154,62 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
 #'  }
 #'
 #' @importFrom htmltools tags
-#' @importFrom shiny reactiveValues observeEvent removeUI updateQueryString insertUI icon
+#' @importFrom shiny reactiveValues observeEvent removeUI updateQueryString insertUI is.reactive icon updateActionButton updateTextInput renderUI
 #' @importFrom stats setNames
-auth_server <- function(input, output, session, check_credentials, use_token = FALSE) {
-
+auth_server <- function(input, output, session, check_credentials, 
+                        use_token = FALSE, lan = NULL) {
+  
   ns <- session$ns
   jns <- function(x) {
     paste0("#", ns(x))
   }
-
-  lan <- use_language()
-
+  
+  if(!is.reactive(lan)){
+    if(is.null(lan)){
+      lan <- reactive(use_language())
+    } else {
+      lan <- reactive(lan)
+    }
+  }
+  
+  
+  observe({
+    session$sendCustomMessage(
+      type = "focus_input",
+      message = list(inputId = ns("user_id"))
+    )
+  })
+  
+  observe({
+    if(!is.null(input$language)){
+      lan()$set_language(input$language) 
+      updateTextInput(session, inputId = "user_id", label = lan()$get("Username:"))
+      updateTextInput(session, inputId = "user_pwd", label = lan()$get("Password:"))
+      updateActionButton(session, inputId = "go_auth", label = lan()$get("Login"))
+      
+      session$sendCustomMessage(
+        type = "update_auth_title",
+        message = list(
+          inputId = ns("shinymanager-auth-head"), 
+          title = lan()$get("Please authenticate")
+        )
+      )
+      
+      output$update_shinymanager_language <- renderUI({
+        shinymanager_language(lan()$get_language())
+      })
+      
+      output$label_language <- renderUI({
+        tags$p(paste0(lan()$get("Language"), " :"), 
+               style = "text-align: right; font-style: italic; margin-top:5px")
+      })
+      
+    }
+  })
+  
+  
   authentication <- reactiveValues(result = FALSE, user = NULL, user_info = NULL)
-
+  
   observeEvent(input$go_auth, {
     removeUI(selector = jns("msg_auth"))
     res_auth <- check_credentials(input$user_id, input$user_pwd)
@@ -136,14 +220,14 @@ auth_server <- function(input, output, session, check_credentials, use_token = F
       authentication$user_info <- res_auth$user_info
       # token <- generate_token(input$user_id)
       token <- .tok$generate(input$user_id)
-
+      
       if (isTRUE(use_token)) {
         # add_token(token, as.list(res_auth$user_info))
         .tok$add(token, as.list(res_auth$user_info))
-        updateQueryString(queryString = paste0("?token=", token), session = session)
+        updateQueryString(queryString = paste0("?token=", token, "&language=", lan()$get_language()), session = session)
         session$reload()
       }
-
+      
     } else {
       if (is.null(res_auth$user_info)) {
         save_logs_failed(input$user_id, status = "Unknown user")
@@ -151,7 +235,7 @@ auth_server <- function(input, output, session, check_credentials, use_token = F
           selector = jns("result_auth"),
           ui = tags$div(
             id = ns("msg_auth"), class = "alert alert-danger",
-            icon("exclamation-triangle"), lan$get("Username or password are incorrect")
+            icon("exclamation-triangle"), lan()$get("Username or password are incorrect")
           )
         )
       } else if (isTRUE(res_auth$expired)) {
@@ -160,7 +244,7 @@ auth_server <- function(input, output, session, check_credentials, use_token = F
           selector = jns("result_auth"),
           ui = tags$div(
             id = ns("msg_auth"), class = "alert alert-danger",
-            icon("exclamation-triangle"), lan$get("Your account has expired")
+            icon("exclamation-triangle"), lan()$get("Your account has expired")
           )
         )
       } else {
@@ -170,7 +254,7 @@ auth_server <- function(input, output, session, check_credentials, use_token = F
             selector = jns("result_auth"),
             ui = tags$div(
               id = ns("msg_auth"), class = "alert alert-danger",
-              icon("exclamation-triangle"), lan$get("You are not authorized for this application")
+              icon("exclamation-triangle"), lan()$get("You are not authorized for this application")
             )
           )
         } else {
@@ -179,14 +263,14 @@ auth_server <- function(input, output, session, check_credentials, use_token = F
             selector = jns("result_auth"),
             ui = tags$div(
               id = ns("msg_auth"), class = "alert alert-danger",
-              icon("exclamation-triangle"), lan$get("Username or password are incorrect")
+              icon("exclamation-triangle"), lan()$get("Username or password are incorrect")
             )
           )
         }
       }
     }
   }, ignoreInit = TRUE)
-
+  
   return(authentication)
 }