+++ inst/assets/bindEnter.js
@@ -23,3 +23,11 @@ function bindEnter(ns) {
     }
   });
 }
+
+Shiny.addCustomMessageHandler('focus_input', function(data) {
+  $('#' + data.inputId).focus();
+});
+
+Shiny.addCustomMessageHandler('update_auth_title', function(data) {
+  $('#' + data.inputId).html(data.title);
+});