+++ NEWS.md
@@ -1,3 +1,15 @@
+# shinymanager 1.0.300
+
+* Add ``autofocus`` on username input.
+* Fix some (strange) bug with ``input$shinymanager_where``
+* Fix `inputs_list` with some shiny version
+* `auth_ui()` now accept a `choose_language` arguments.
+* Rename `br` language into `pt-BR` (iso code)
+* add user info in downloaded log file
+* add `set_labels()` for customize labels
+* Fix simultaneous admin session
+* (#37) hashing password using `scrypt`
+
 # shinymanager 1.0.200
 
 * Can configure shiny input for editing user in `secure_app()` with new `inputs_list`.